+++ apps/docs/src/app/core/component-docs/slider/examples/ticks-and-labels/slider-ticks-and-labels-example.component.ts
@@ -1,0 +1,11 @@
+import { ChangeDetectionStrategy, Component } from '@angular/core';
+
+@Component({
+    selector: 'fd-slider-ticks-and-labels-example',
+    templateUrl: './slider-ticks-and-labels-example.component.html',
+    changeDetection: ChangeDetectionStrategy.OnPush
+})
+export class SliderTicksAndLabelsExampleComponent {
+    value = 25;
+    value1 = 25;
+}