+++ apps/docs/src/app/platform/component-docs/platform-forms/multi-input/platform-multi-input-docs.module.ts
@@ -15,24 +15,24 @@ import {
 import { DialogModule, InputGroupModule, MOBILE_MODE_CONFIG } from '@fundamental-ngx/core';
 
 import { PlatformMultiInputDocsComponent } from './platform-multi-input-docs.component';
-import { PlatformMulitInputExampleComponent } from './platform-mulit-input-example/platform-mulit-input-example.component';
-import { PlatformMulitInputHeaderComponent } from './platform-mulit-input-header/platform-mulit-input-header.component';
-import { PlatformMulitInputGroupedExampleComponent } from './platform-mulit-input-example/platform-mulit-input-grouped-example.component';
-import { PlatformMulitInputDeclineExampleComponent } from './platform-mulit-input-example/platform-mulit-input-decline-example.component';
-import { PlatformMulitInputComplexExampleComponent } from './platform-mulit-input-example/platform-mulit-input-complex-example.component';
-import { PlatformMulitInputMobileExampleComponent } from './platform-mulit-input-example/platform-mulit-input-mobile-example.component';
+import { PlatformMultiInputExampleComponent } from './platform-multi-input-example/platform-multi-input-example.component';
+import { PlatformMultiInputHeaderComponent } from './platform-multi-input-header/platform-multi-input-header.component';
+import { PlatformMultiInputGroupedExampleComponent } from './platform-multi-input-example/platform-multi-input-grouped-example.component';
+import { PlatformMultiInputDeclineExampleComponent } from './platform-multi-input-example/platform-multi-input-decline-example.component';
+import { PlatformMultiInputComplexExampleComponent } from './platform-multi-input-example/platform-multi-input-complex-example.component';
+import { PlatformMultiInputMobileExampleComponent } from './platform-multi-input-example/platform-multi-input-mobile-example.component';
 import { ApiComponent } from '../../../../documentation/core-helpers/api/api.component';
 import { SharedDocumentationPageModule } from '../../../../documentation/shared-documentation-page.module';
 import { API_FILES } from '../../../api-files';
-import { PlatformMulitInputDisabledExampleComponent } from './platform-mulit-input-example/platform-mulit-input-disabled.component';
+import { PlatformMultiInputDisabledExampleComponent } from './platform-multi-input-example/platform-multi-input-disabled.component';
 import { SharedDocumentationModule } from '../../../../documentation/shared-documentation.module';
 import { MULTI_INPUT_MOBILE_CONFIG } from '../../../../documentation/utilities/consts/mobile-mode-configuration-tokens';
-import { PlatformMulitInputReactiveExampleComponent } from './platform-mulit-input-example/platform-mulit-input-reactive-example.component';
+import { PlatformMultiInputReactiveExampleComponent } from './platform-multi-input-example/platform-multi-input-reactive-example.component';
 
 const routes: Routes = [
     {
         path: '',
-        component: PlatformMulitInputHeaderComponent,
+        component: PlatformMultiInputHeaderComponent,
         children: [
             { path: '', component: PlatformMultiInputDocsComponent },
             { path: 'api', component: ApiComponent, data: { content: API_FILES.multiInput } }
@@ -43,15 +43,15 @@ const routes: Routes = [
 @NgModule({
     declarations: [
         PlatformMultiInputDocsComponent,
-        PlatformMulitInputExampleComponent,
-        PlatformMulitInputHeaderComponent,
-        PlatformMulitInputGroupedExampleComponent,
-        PlatformMulitInputComplexExampleComponent,
-        PlatformMulitInputDeclineExampleComponent,
-        PlatformMulitInputDeclineExampleComponent,
-        PlatformMulitInputDisabledExampleComponent,
-        PlatformMulitInputMobileExampleComponent,
-        PlatformMulitInputReactiveExampleComponent
+        PlatformMultiInputExampleComponent,
+        PlatformMultiInputHeaderComponent,
+        PlatformMultiInputGroupedExampleComponent,
+        PlatformMultiInputComplexExampleComponent,
+        PlatformMultiInputDeclineExampleComponent,
+        PlatformMultiInputDeclineExampleComponent,
+        PlatformMultiInputDisabledExampleComponent,
+        PlatformMultiInputMobileExampleComponent,
+        PlatformMultiInputReactiveExampleComponent
     ],
     imports: [
         CommonModule,