+++ apps/docs/src/app/platform/component-docs/platform-standard-list-item/platform-standard-list-item-examples/platform-standard-list-item-with-group-header-example.component.ts
@@ -1,7 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-    selector: 'fdp-standard-list-item-with-group-header-example',
-    templateUrl: './platform-standard-list-item-with-group-header-example.component.html'
-})
-export class PlatformStandardListItemtWithGroupHeaderExampleComponent { }