+++ apps/docs/src/app/platform/component-docs/platform-standard-list-item/platform-standard-list-item-examples/platform-standard-list-item-with-selection-example.component.ts
@@ -55,7 +55,7 @@ export class ListDataProvider extends DataProvider<Address> {
     }
 }
 @Component({
-    selector: 'fdp-standard-list-item-with-selection-example',
+    selector: 'fdp-platform-standard-list-item-with-selection-example',
     templateUrl: './platform-standard-list-item-with-selection-example.component.html'
 })
 export class PlatformStandardListItemWithSelectionExampleComponent implements AfterViewInit {