+++ libs/core/src/lib/slider/public_api.ts
@@ -1,0 +1,3 @@
+export * from './slider.module';
+export * from './slider.component';
+export * from './slider.model';